Rename AnclapInfoResponse to AnclapTransferServerInfoResponse
